Separate the GenerateRandomAlphaOnlyString function from stringutils
